+++ test.js
@@ -1,21 +1,30 @@
 import postcss from 'postcss';
 import test    from 'ava';
+import fs      from 'fs';
 
 import plugin from './';
 
 function run(t, input, output, opts = { }) {
     return postcss([ plugin(opts) ]).process(input)
         .then( result => {
-            t.same(result.css, output);
-            t.same(result.warnings().length, 0);
+            t.deepEqual(result.css, output);
+            t.deepEqual(result.warnings().length, 0);
+        });
+}
+
+function runFile(t, input, output, opts = { }) {
+    return postcss([ plugin(opts) ]).process(fs.readFileSync(input), { from: input })
+        .then( result => {
+            t.deepEqual(result.css, output);
+            t.deepEqual(result.warnings().length, 0);
         });
 }
 
 function runWithWarn(t, input, output, opts = { }) {
     return postcss([ plugin(opts) ]).process(input)
         .then( result => {
-            t.same(result.css, output);
-            t.same(result.warnings().length, 1);
+            t.deepEqual(result.css, output);
+            t.deepEqual(result.warnings().length, 1);
         });
 }
 
@@ -25,13 +34,13 @@ test('Always pass', t => {
 });
 
 test('LightMuted twice', t => {
-    return run(t, 'a { background: get-color("images/2.jpg", LightMuted); ' +
-                  'color: get-color("images/white.jpg", LightMuted); }',
+    return run(t, 'a { background: get-color("test/car.jpg", LightMuted); ' +
+                  'color: get-color("test/white.jpg", LightMuted); }',
                   'a { background: #e1e1e1; color: #fcfcfc; }', { });
 });
 
 test('Vibrant by arg', t => {
-    return run(t, 'a { background: get-color("images/girl.png", Vibrant); }',
+    return run(t, 'a { background: get-color("test/img/girl.png", Vibrant); }',
                   'a { background: #e8ba3c; }', { });
 });
 
@@ -42,39 +51,54 @@ test('Vibrant by Url', t => {
 });
 
 test('Title color', t => {
-    return run(t, 'a { background: get-color("images/2.jpg", Vibrant, title); }',
+    return run(t, 'a { background: get-color("test/car.jpg", Vibrant, title); }',
                   'a { background: #fff; }', { });
 });
 
 test('Body color with warning', t => {
-    return runWithWarn(t, 'a { background: get-color("images/white.jpg", LightVibrant, body); }',
+    return runWithWarn(t, 'a { background: get-color("test/white.jpg", LightVibrant, body); }',
                   'a { background: #000; }', { });
 });
 
 test('Body color without warning', t => {
-    return run(t, 'a { background: get-color("images/white.jpg", LightMuted, body); }',
+    return run(t, 'a { background: get-color("test/white.jpg", LightMuted, body); }',
                   'a { background: #000; }', { });
 });
 
 test('Vibrant by default', t => {
-    return run(t, 'a { background: get-color("images/2.jpg"); }',
-                  'a { background: #b9911b; }', { });
+    return run(t, 'a { background: get-color("test/car.jpg"); }',
+                  'a { background: #b79022; }', { });
 });
 
 test('Works with single quotes', t => {
-    return run(t, 'a { background: get-color(\'images/2.jpg\'); }',
-                  'a { background: #b9911b; }', { });
+    return run(t, 'a { background: get-color(\'test/car.jpg\'); }',
+                  'a { background: #b79022; }', { });
 });
 
 test('White image take LighMuted by default', t => {
-    return runWithWarn(t, 'a { background: get-color("images/white.jpg")' +
-                       ' url(\'2.jpg\') no-repeat; }',
-                  'a { background: #fcfcfc url(\'2.jpg\') no-repeat; }', { });
+    return runWithWarn(t, 'a { background: get-color("test/white.jpg")' +
+                       ' url("car.jpg") no-repeat; }',
+                  'a { background: #fcfcfc url("car.jpg") no-repeat; }', { });
 });
 
 test('Multiple value', t => {
-    return runWithWarn(t, 'a { background: get-color("images/white.jpg")' +
-                       ' url(\'2.jpg\') no-repeat; }',
-                  'a { background: #fcfcfc url(\'2.jpg\') no-repeat; }', { });
+    return runWithWarn(t, 'a { background: get-color("test/white.jpg")' +
+                       ' url("car.jpg") no-repeat; }',
+                  'a { background: #fcfcfc url("car.jpg") no-repeat; }', { });
+});
+
+test('Relative path in file', t => {
+    return runFile(t, './test/css/sample.css',
+                  '.bg { background: url("../img/girl.jpg") #e8ba3c; }\n', { });
+});
+
+test('Url in file', t => {
+    return runFile(t, './test/css/url.css',
+                  '.url { color: #b9911b; }\n', { });
+});
+
+test('Relative path for deep file', t => {
+    return runFile(t, './test/css/deep/sample.css',
+                  '.bg { background: url("../../img/girl.jpg") #e8ba3c; }\n', { });
 });