+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## v0.2.0 - 20170929
+
+- Improve warning message when color is missing
+- Upgrade to PostCSS v6.0.11
+- Upgrade to AVA v0.22.0
+- Fix image path to be relative [#2](https://github.com/ismamz/postcss-get-color/issues/2)
+- Fix some typo issues
+- Add `test` folder and compress sample images
+
 ## v0.1.0 - 20160412
 
 - Initial release